+++ TradingApp/Models/ValuationResult.cs
@@ -3,7 +3,7 @@
     public class ValuationResult
     {
         public long TradeId { get; set; }
-        public decimal PV { get; set; }
+        public double PV { get; set; }
         public string ErrorMessage { get; set; }
     }
 }
